+++ src/main/java/io/github/Victor5H/dao/EmployeeDAO.java
@@ -7,9 +7,9 @@ import io.github.Victor5H.entities.Employee;
 public interface EmployeeDAO {
 	boolean insertEmp(Employee e);
 
-	boolean deleteEmp(int id);
+	boolean deleteEmp(Long id);
 
-	Employee getEmp(int id);
+	Employee getEmp(Long id);
 
 	List<Employee> getAllEmp();
